requirements.txt
langchain
llama-cpp-python
openai
python-dotenv
